server/MyNews/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
dj-database-url==0.5.0
Django==3.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4
python-decouple==3.2
pytz==2020.1
sqlparse==0.3.1
whitenoise==4.1.4

requests~=2.24.0
bs4~=4.9.1
beautifulsoup4~=4.9.1